Add files via upload
Here is a new solution I created. It is much faster than the old one as it is only O(n^2), not O(n^3). It can generate answers up to about n=40 000 within 30 minutes, but we need to get it to n=1 000 000. I believe this could be solved by using a greedy algorithm to generate the initial board.